Funciona punto 2.F

+++ demo/src/main/java/com/example/demo/dtos/DtoCarrera.java
@@ -5,10 +5,11 @@ import lombok.Getter;
 @Getter
 public class DtoCarrera {
 	private String nombre;
-	private int Cant;
+	private Long Cant;
 
 	public DtoCarrera(Object resultado, Object resultado2) {
 		this.nombre = (String) resultado;
-		this.Cant = (int) resultado2;
+		this.Cant = (Long) resultado2;
 	}
 }
+
